requirements.txt
aiodns==3.2.0
aiofiles==23.2.1
aiohappyeyeballs==2.3.6
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.3
black==24.4.2
Brotli==1.1.0
brotlicffi==1.1.0.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
defusedxml==0.8.0rc2
discord.py==2.4.0
flake8==7.1.1
fonttools==4.53.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imutils==0.5.4
kiwisolver==1.4.5
matplotlib==3.9.1
mccabe==0.7.0
multidict==6.0.5
mypy_extensions==1.0.0
numpy==1.26.4
olefile==0.47
ollama==0.3.3
opencv @ file:///build/source/modules/python/package/dist/opencv-4.9.0-py3-none-any.whl#sha256=0b1c6093b66ec745df186877df1e6e72241dfe8dda392f023bc361fe1d5411ca
packaging==24.1
paramiko==3.4.0
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
proxmoxer==2.1.0
pyasn1==0.6.0
pycares==4.4.0
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
PyNaCl==1.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyvips==2.2.1
requests==2.32.3
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
